Changed logic and the color of basic mobs
Changed the color of basic mobs to black for ease of reading, as well as changed the text on the item to not be italicized and formatted the entity types properly.  Also cut down on how much I would regrab something like an item or entity's type / hash map objects.